+++ eNMS/models/services/file_transfer/netmiko_file_transfer.py
@@ -1,5 +1,5 @@
 from netmiko import file_transfer
-from sqlalchemy import Boolean, ForeignKey, Integer
+from sqlalchemy import Boolean, Float, ForeignKey, Integer
 from wtforms.validators import InputRequired
 
 from eNMS.database import db
@@ -21,6 +21,11 @@ class NetmikoFileTransferService(ConnectionService):
     direction = db.Column(db.SmallString)
     disable_md5 = db.Column(Boolean, default=False)
     driver = db.Column(db.SmallString)
+    conn_timeout = db.Column(Float, default=10.0)
+    auth_timeout = db.Column(Float, default=0.0)
+    banner_timeout = db.Column(Float, default=15.0)
+    fast_cli = db.Column(Boolean, default=False)
+    global_delay_factor = db.Column(Float, default=1.0)
     file_system = db.Column(db.SmallString)
     inline_transfer = db.Column(Boolean, default=False)
     overwrite_file = db.Column(Boolean, default=False)
@@ -31,9 +36,12 @@ class NetmikoFileTransferService(ConnectionService):
         netmiko_connection = run.netmiko_connection(device)
         source = run.sub(run.source_file, locals())
         destination = run.sub(run.destination_file, locals())
-        if str(vs.file_path) not in source:
+        if run.direction == "put" and str(vs.file_path) not in source:
             source = f"{vs.file_path}{source}"
+        if run.direction == "get" and str(vs.file_path) not in destination:
+            destination = f"{vs.file_path}{destination}"
         run.log("info", f"Transferring file {source}", device)
+        netmiko_connection.password = run.get_credentials(device).get("password")
         transfer_dict = file_transfer(
             netmiko_connection,
             source_file=source,
@@ -44,6 +52,7 @@ class NetmikoFileTransferService(ConnectionService):
             disable_md5=run.disable_md5,
             inline_transfer=run.inline_transfer,
         )
+        netmiko_connection.password = "*" * 8
         return {"success": True, "result": transfer_dict}