+++ eNMS/static/js/administration.js
@@ -241,7 +241,7 @@ function scanCluster() {
   });
 }
 
-function editFile(filename, filepath) {
+function editFile(id, filename, filepath) {
   call({
     url: `/edit_file/${filename}`,
     callback: function(content) {
@@ -252,9 +252,9 @@ function editFile(filename, filepath) {
       openPanel({
         name: "file_editor",
         title: `Edit ${filepath}`,
-        id: filename,
+        id: id,
         callback: () => {
-          const display = document.getElementById(`file_content-${filename}`);
+          const display = document.getElementById(`file_content-${id}`);
           // eslint-disable-next-line new-cap
           let fileEditor = (editors[filename] = CodeMirror.fromTextArea(display, {
             lineWrapping: true,
@@ -287,14 +287,14 @@ function saveFile(file) {
 
 function showFileUploadPanel(folder) {
   if (!folder) folder = folderPath;
-  const path = folder.replace(/\//g, ">");
+  const pathId = folder.replace(/\//g, ">") || 1;
   openPanel({
     name: "upload_files",
     title: `Upload files to ${folder}`,
     size: "700 615",
-    id: path,
+    id: pathId,
     callback: () => {
-      const element = document.getElementById(`dropzone-${path}`);
+      const element = document.getElementById(`dropzone-${pathId}`);
       let dropzone = new Dropzone(element, {
         url: "/upload_files",
         autoProcessQueue: false,
@@ -315,8 +315,8 @@ function showFileUploadPanel(folder) {
         },
         timeout: settings.files.upload_timeout,
       });
-      $(`[id="dropzone-submit-${path}"]`).click(function() {
-        $(`[id="folder-${path}"]`).val(folder.slice(6));
+      $(`[id="dropzone-submit-${pathId}"]`).click(function() {
+        $(`[id="folder-${pathId}"]`).val(folder);
         dropzone.processQueue();
       });
     },